Social_App/requirement.txt
fastapi[all]
python-dotenv
fastapi-users[sqlalchemy]
imagekitio
uvicorn[standard]
aiosqlite
psycopg2
asyncpg
streamlit
requests